+++ exporter/exporter_test.go
@@ -116,7 +116,7 @@ const (
         redfish_thermal_sensor_temperature{chassisModel="model a",chassisSerialNumber="SN98765",name="01-Inlet Ambient"} 22
 	`
 	GoodPowerVoltageOutputExpected = `
-        # HELP redfish_power_voltage_output Power voltage output in watts
+        # HELP redfish_power_voltage_output Power voltage output in volts
         # TYPE redfish_power_voltage_output gauge
         redfish_power_voltage_output{chassisModel="model a",chassisSerialNumber="SN98765",name="PSU1_VOUT"} 12.2
 	`
@@ -126,7 +126,7 @@ const (
         redfish_power_voltage_status{chassisModel="model a",chassisSerialNumber="SN98765",name="PSU1_VOUT"} 1
 	`
 	GoodPowerSupplyOutputOemHpeExpected = `
-        # HELP redfish_power_supply_output Power supply output in watts
+        # HELP redfish_power_supply_output The average power output of Power Supply in watts
         # TYPE redfish_power_supply_output gauge
         redfish_power_supply_output{bayNumber="1",chassisModel="model a",chassisSerialNumber="SN98765",firmwareVersion="x.xx",manufacturer="DELTA",model="psmodel",name="HpeServerPowerSupply",powerSupplyType="AC",serialNumber="999999"} 91
 	`
@@ -136,7 +136,7 @@ const (
         redfish_power_supply_status{bayNumber="1",chassisModel="model a",chassisSerialNumber="SN98765",firmwareVersion="x.xx",manufacturer="DELTA",model="psmodel",name="HpeServerPowerSupply",powerSupplyType="AC",serialNumber="999999"} 1
 	`
 	GoodPowerSupplyOutputOemHpExpected = `
-        # HELP redfish_power_supply_output Power supply output in watts
+        # HELP redfish_power_supply_output The average power output of Power Supply in watts
         # TYPE redfish_power_supply_output gauge
         redfish_power_supply_output{bayNumber="2",chassisModel="model a",chassisSerialNumber="SN98765",firmwareVersion="x.xx",manufacturer="DELTA",model="psmodel",name="HpeServerPowerSupply",powerSupplyType="AC",serialNumber="999999"} 91
 	`